+++ src/main/java/org/opensearch/security/dlic/rest/api/SecurityRestApiActions.java
@@ -26,7 +26,8 @@ import org.opensearch.security.configuration.ConfigurationRepository;
 import org.opensearch.security.configuration.SecurityConfigVersionHandler;
 import org.opensearch.security.configuration.SecurityConfigVersionsLoader;
 import org.opensearch.security.hasher.PasswordHasher;
-import org.opensearch.security.privileges.PrivilegesEvaluator;
+import org.opensearch.security.privileges.PrivilegesConfiguration;
+import org.opensearch.security.privileges.RoleMapper;
 import org.opensearch.security.resources.ResourcePluginInfo;
 import org.opensearch.security.resources.ResourceSharingIndexHandler;
 import org.opensearch.security.resources.api.migrate.MigrateResourceSharingInfoApiAction;
@@ -49,7 +50,8 @@ public class SecurityRestApiActions {
         final ConfigurationRepository configurationRepository,
         final ClusterService clusterService,
         final PrincipalExtractor principalExtractor,
-        final PrivilegesEvaluator evaluator,
+        final RoleMapper roleMapper,
+        final PrivilegesConfiguration privilegesConfiguration,
         final ThreadPool threadPool,
         final AuditLog auditLog,
         final SslSettingsManager sslSettingsManager,
@@ -62,11 +64,11 @@ public class SecurityRestApiActions {
         final var securityApiDependencies = new SecurityApiDependencies(
             adminDns,
             configurationRepository,
-            evaluator,
-            new RestApiPrivilegesEvaluator(settings, adminDns, evaluator, principalExtractor, configPath, threadPool),
+            privilegesConfiguration,
+            new RestApiPrivilegesEvaluator(settings, adminDns, roleMapper, principalExtractor, configPath, threadPool),
             new RestApiAdminPrivilegesEvaluator(
                 threadPool.getThreadContext(),
-                evaluator,
+                privilegesConfiguration,
                 adminDns,
                 settings.getAsBoolean(SECURITY_RESTAPI_ADMIN_ENABLED, false)
             ),
@@ -91,7 +93,7 @@ public class SecurityRestApiActions {
                     configurationRepository,
                     clusterService,
                     principalExtractor,
-                    evaluator,
+                    roleMapper,
                     threadPool,
                     auditLog
                 ),