+++ src/test/java/org/opensearch/security/filter/SecurityFilterTests.java
@@ -30,7 +30,7 @@ import org.opensearch.security.configuration.ClusterInfoHolder;
 import org.opensearch.security.configuration.CompatConfig;
 import org.opensearch.security.configuration.DlsFlsRequestValve;
 import org.opensearch.security.http.XFFResolver;
-import org.opensearch.security.privileges.PrivilegesEvaluator;
+import org.opensearch.security.privileges.PrivilegesConfiguration;
 import org.opensearch.security.privileges.ResourceAccessEvaluator;
 import org.opensearch.security.resolver.IndexResolverReplacer;
 import org.opensearch.security.support.ConfigConstants;
@@ -80,7 +80,7 @@ public class SecurityFilterTests {
     public void testImmutableIndicesWildcardMatcher() {
         final SecurityFilter filter = new SecurityFilter(
             settings,
-            mock(PrivilegesEvaluator.class),
+            mock(PrivilegesConfiguration.class),
             mock(AdminDNs.class),
             mock(DlsFlsRequestValve.class),
             mock(AuditLog.class),
@@ -105,7 +105,7 @@ public class SecurityFilterTests {
 
         final SecurityFilter filter = new SecurityFilter(
             settings,
-            mock(PrivilegesEvaluator.class),
+            mock(PrivilegesConfiguration.class),
             mock(AdminDNs.class),
             mock(DlsFlsRequestValve.class),
             auditLog,